requirements.txt
# Core LangGraph / LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# Search
tavily-python>=0.5.0

# Data & Config
pydantic>=2.0.0
python-dotenv>=1.0.0

# Logging & CLI
rich>=13.0.0
typer>=0.12.0

# Optional: Observability
langsmith>=0.1.0

# Dev dependencies
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.5.0
mypy>=1.10.0
